+++ docs/_pages/changelog.markdown
@@ -6,5 +6,6 @@ permalink: changelog/
 
 This page tracks major changes to the site, in reverse chronological order:
 
+- Generating basic scatter plots to visualize alt. minority ratings
 - Bound ag-grid download to v30, to fix ratings tables
 - First snapshot